Create generate_unit_list.F90
  Created the new file generate_unit_list.F90, which can be compiled to give the
stand-alone tool generate_unit_list.  This extacts the rescaled unit
descriptions in the syntax like [L T-1 ~> m s-1] from the file given by a
command line argument, and returns a list of the various units that have
distinct dimensional scaling and the counts of their occurance, accoording to a
key of rescaled variables that can also be provided as a run-time parameter.

  The syntax for generate_unit_list is:

  generate_unit_list [-k <key>] [--simplify] [-f] <input_file>

<input_file> is the name of the file to parse for the unit descriptions.

<key> is the key of rescaled units in the double-quote delimited syntax, but
  if omitted the key defaults to the full list currently used by MOM6:
  '["Z", "H", "L", "T", "R", "Q", "C", "S"]'

--simplify specifies that all units with common rescaling are collapsed to a
  single entry giving the most frequently used unit description.  If it is
  omitted all unique unit descriptions are listed, but grouped by common
  rescaling and with added lines separating groups with different rescaling to
  aid in the manual consolidation of descriptions with common rescaling.